Initial Prototype for WASI Preview 1
Also moved WASIp2 implementation to separate file.
