+++ crates/wasi-isolated-fs/Cargo.toml
@@ -21,6 +21,7 @@ bytes = { workspace = true }
 memchr = { workspace = true }
 camino = { workspace = true }
 rand_xoshiro = { workspace = true }
+wiggle = { workspace = true }
 
 [dependencies.wasmtime]
 workspace = true
